+++ dwara-misc/dwara-directorywatcher-standalone/src/main/java/org/ishafoundation/dwara/misc/watcher/DirectoryWatcherForCopiedFiles.java
@@ -31,14 +31,20 @@ import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
 import java.util.Map.Entry;
+import java.util.concurrent.Executor;
+import java.util.concurrent.LinkedBlockingQueue;
+import java.util.concurrent.ThreadPoolExecutor;
 import java.util.concurrent.TimeUnit;
 
 import javax.xml.bind.DatatypeConverter;
 
 import org.apache.commons.io.FileUtils;
 import org.apache.commons.lang3.StringUtils;
+import org.ishafoundation.dwaraapi.commandline.local.CommandLineExecuterImpl;
 import org.ishafoundation.dwaraapi.enumreferences.Checksumtype;
 import org.ishafoundation.dwaraapi.utils.ChecksumUtil;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 
 /**
@@ -47,6 +53,8 @@ import org.ishafoundation.dwaraapi.utils.ChecksumUtil;
 
 public class DirectoryWatcherForCopiedFiles {
 
+	private static Logger logger = LoggerFactory.getLogger(DirectoryWatcherForCopiedFiles.class);
+	
 	private final WatchService watchService;
 	private final Path watchedDir;
 	private final Map<WatchKey,Path> keys;
@@ -55,6 +63,7 @@ public class DirectoryWatcherForCopiedFiles {
 	private Long newFileWait = 10000L;
 	private static Path opsDirLoc = null;
 	private static List<String> csvEntries = null;
+	private static Executor executor = null;
 	private static File csvFile = null;
 
 	@SuppressWarnings("unchecked")
@@ -160,6 +169,7 @@ public class DirectoryWatcherForCopiedFiles {
 						if (kind == ENTRY_CREATE) {
 							try {
 //								createCSVEntry(child);
+								updateStatus(child, Status.copying);
 //								Path prev = keys.get(key);
 //								if (prev != null && !dir.equals(prev)) {
 //									updateStatus(child, Status.done);
@@ -171,24 +181,25 @@ public class DirectoryWatcherForCopiedFiles {
 									registerAll(child);
 								}
 								
-								Path newDir = keys.get(key);
-								System.out.format("test :: %s -> %s\n", newDir, child);
-								if (newDir != null && !dir.equals(newDir)) { // if its renamed
-									renameEntry(newDir, child);
-									System.out.format("update: %s -> %s\n", newDir, child);
-								}
-								else {
-									createCSVEntry(child);
-								}
+//								Path newDir = keys.get(key);
+//								//System.out.format("test :: %s -> %s\n", newDir, child);
+//								if (newDir != null && !dir.equals(newDir)) { // if its renamed
+//									renameEntry(newDir, child);
+//									System.out.format("update: %s -> %s\n", newDir, child);
+//								}
+//								else {
+//									createCSVEntry(child);
+//								}
 
 							} catch (IOException x) {
 								// ignore to keep sample readbale
 							}
 						}
 						else if (kind == ENTRY_MODIFY) {
-							Path artifactPath = getArtifactPath(child);
-							if (trace)
-								System.out.println("artifactPath " + artifactPath);
+//							Path artifactPath = getArtifactPath(child);
+//							if (trace)
+//								System.out.println("artifactPath " + artifactPath);
+							
 							// Update modified time
 							BasicFileAttributes attrs = null;
 							try {
@@ -200,12 +211,12 @@ public class DirectoryWatcherForCopiedFiles {
 							}
 							FileTime lastModified = attrs.lastModifiedTime();
 							if (trace)
-								System.out.println("lastModified " + lastModified);
-							expirationTimes.put(artifactPath, lastModified.toMillis()+newFileWait);
+								System.out.println(child + " lastModified " + lastModified);
+							expirationTimes.put(child, lastModified.toMillis()+newFileWait);
 						}
-						else if (kind == ENTRY_DELETE) {
-							updateStatus(child, Status.moved);
-						}
+//						else if (kind == ENTRY_DELETE) {
+//							updateStatus(child, Status.moved);
+//						}
 					}
 
 					// reset key and remove from set if directory no longer accessible
@@ -228,7 +239,7 @@ public class DirectoryWatcherForCopiedFiles {
 						System.out.println("No pending events");
 				}
 
-				handleExpiredWaitTimes(currentTime);
+				handleExpiredWaitTimes(key, currentTime);
 
 				// If there are no files left stop polling and block on .take()
 				if(expirationTimes.isEmpty())
@@ -240,26 +251,6 @@ public class DirectoryWatcherForCopiedFiles {
 					System.out.println("timeout: "+timeout);
 				key = watchService.poll(timeout, TimeUnit.MILLISECONDS);
 			}
-		}
-	}
-
-
-	private void moveFolderToOpsAreaAndOrganise(Path srcPath) throws Exception {
-		try {
-			String artifactName = getArtifactName(srcPath);
-			final Path dest = Paths.get(opsDirLoc.toString(), artifactName, "mxf");
-			Files.createDirectories(dest);
-			Files.walkFileTree(srcPath, new SimpleFileVisitor<Path>() {
-				@Override
-				public FileVisitResult visitFile(Path file, BasicFileAttributes attrs) throws IOException {
-					Files.move(file, Paths.get(dest.toString(), file.getFileName().toString()), StandardCopyOption.ATOMIC_MOVE);
-					return FileVisitResult.CONTINUE;
-				}
-			});
-			Files.delete(srcPath);
-		} catch (IOException e) {
-			e.printStackTrace();
-			updateStatus(srcPath, Status.move_failed);
 		}
 	}
 
@@ -276,60 +267,168 @@ public class DirectoryWatcherForCopiedFiles {
 		return artifactName;
 	}
 	
-	private void handleExpiredWaitTimes(Long currentTime) throws Exception {
+	private void handleExpiredWaitTimes(WatchKey watchKey, Long currentTime) throws Exception {
 		// Start import for files for which the expirationtime has passed
-
 		Iterator<Entry<Path, Long>> it = expirationTimes.entrySet().iterator();
 		while (it.hasNext())
 		{
 			Entry<Path, Long> item = (Entry<Path, Long>) it.next();
 			Long expiryTime = (Long) item.getValue();
+			Path filePath = item.getKey();
+			
 			if(expiryTime <= currentTime) {
-				action(item.getKey());
+				if(filePath.getFileName().toString().endsWith(".mxf"))
+					verifyChecksumAndMoveToOpsArea(watchKey, filePath);
+				
+//				Integer copiedFileCountOfArtifact = artifactPath_CopiedFileCount.get(artifactPath);
+//				//System.out.println(key + ":" + copiedFileCountOfArtifact);
+//				if(copiedFileCountOfArtifact != null)
+//					copiedFileCountOfArtifact = copiedFileCountOfArtifact + 1;
+//				else 
+//					copiedFileCountOfArtifact = 1;
+//				
+//				artifactPath_CopiedFileCount.put(artifactPath, copiedFileCountOfArtifact);
+//				System.out.println(artifactPath + ":" + copiedFileCountOfArtifact);
+//				if(copiedFileCountOfArtifact == 1) {
+//					System.out.println(artifactPath + " copy completed");
+//					verifyChecksumAndMoveToOpsArea(artifactPath);
+//				}
 				// System.out.println("expired " + item.getKey());
-				// do something with the file
+
 				it.remove();
 			}
 		}
 	}
 
-    private void action(Path child) throws Exception{
-		System.out.println(child + " copy completed");
-		//System.out.println("Method assumes Prasad supplies a \"MD5.txt\" and the video file ending with \".mxf\" in the supplied folder");
-		String expectedMd5 = null;
-		String actualMd5 = null;
-		
-		try {
-			Collection<File> files = FileUtils.listFiles(child.toFile(), null, true);
-			for (File file : files) {
-				String fileName = file.getName();
-				//System.out.println(fileName);
-				if(fileName.endsWith(".md5")) {
-					// TODO : What will be the md5 file content sample looking like ? // d44f11c6df36d8be680db44e969e7ff0  sample-ntsc.mxf
-					expectedMd5 = StringUtils.substringBefore(FileUtils.readFileToString(file), "  ").trim().toUpperCase();
-					System.out.println("expectedMd5 " + expectedMd5);
-				}
-				else if(fileName.endsWith(".mxf")) {
-					updateStatus(child, Status.verifying);
-				    byte[] digest = ChecksumUtil.getChecksum(file, Checksumtype.md5);
-				    actualMd5 = DatatypeConverter.printHexBinary(digest).toUpperCase();
-				    System.out.println("actualMd5 " + actualMd5);
-				}
-			}
-			if(expectedMd5.equals(actualMd5)) {
-				System.out.println("MD5 expected = actual, call dwara api for " + child);
-				updateStatus(child, Status.done);
-				moveFolderToOpsAreaAndOrganise(child);
-			}else {
-				System.out.println(child + "MD5 expected != actual, Now what??? ");
-				updateStatus(child, Status.md5_mismatch);
-			}
-		}catch (Exception e) {
-			e.printStackTrace();
-			updateStatus(child, Status.move_failed);
-		}
+    private void verifyChecksumAndMoveToOpsArea(WatchKey watchKey, Path path) throws Exception{
+		VerifyChecksumAndMoveToOpsAreaTask checksumTask = new VerifyChecksumAndMoveToOpsAreaTask();//applicationContext.getBean(TapeTask.class); 
+		checksumTask.setWatchKey(watchKey);
+		checksumTask.setMxfFilePath(path);
+		executor.execute(checksumTask);
     }
     
+	public class VerifyChecksumAndMoveToOpsAreaTask implements Runnable{
+		
+		private Path mxfFilePathname;
+		
+		private WatchKey watchKey;
+		
+		public void setMxfFilePath(Path mxfFilePathname) {
+			this.mxfFilePathname = mxfFilePathname;
+		}
+		
+		public void setWatchKey(WatchKey watchKey) {
+			this.watchKey = watchKey;
+		}
+		
+		@Override
+		public void run() {
+			String expectedMd5 = null;
+			String actualMd5 = null;
+			Path artifactPath = mxfFilePathname.getParent();
+			File artifactFileObj = artifactPath.toFile();
+			while(true) {
+				//logger.debug(mxfFilePathname + ":" + artifactPath);
+				if(!artifactFileObj.isDirectory())
+					return;
+				Collection<File> files = FileUtils.listFiles(artifactFileObj, null, false);
+				if(files.size() == 4) {
+					try {
+						updateStatus(artifactPath, Status.copy_complete);
+						for (File file : files) {
+							String fileName = file.getName();
+							//System.out.println(fileName);
+							if(fileName.endsWith(".md5")) {
+								// TODO : What will be the md5 file content sample looking like ? // d44f11c6df36d8be680db44e969e7ff0  sample-ntsc.mxf
+								expectedMd5 = StringUtils.substringBefore(FileUtils.readFileToString(file), "  ").trim().toUpperCase();
+							}
+							else if(fileName.endsWith(".mxf")) {
+								updateStatus(artifactPath, Status.verifying);
+								
+							    byte[] digest =  ChecksumUtil.getChecksum(file, Checksumtype.md5);
+							    actualMd5 = DatatypeConverter.printHexBinary(digest).toUpperCase();
+							}
+						}
+						logger.debug(artifactPath + " expectedMd5 " + expectedMd5 + " actualMd5 " + actualMd5);
+						if(expectedMd5.equals(actualMd5)) {
+							updateStatus(artifactPath, Status.verified);
+							moveFolderToOpsAreaAndOrganise(watchKey, artifactPath);
+							updateStatus(artifactPath, Status.moved);
+						}else {
+							logger.error(artifactPath + "MD5 expected != actual, Now what??? ");
+							updateStatus(artifactPath, Status.md5_mismatch);
+						}
+						break;
+					}catch (Exception e) {
+						e.printStackTrace();
+						try {
+							updateStatus(artifactPath, Status.move_failed);
+						} catch (Exception e1) {
+							// TODO Auto-generated catch block
+							e1.printStackTrace();
+						}
+					}
+				}
+				else {
+					try {
+						//System.out.println("waiting for all files");
+						Thread.sleep(1000);
+					} catch (InterruptedException e) {
+						// TODO Auto-generated catch block
+						e.printStackTrace();
+					}
+				}
+			}
+		}
+		
+		private void moveFolderToOpsAreaAndOrganise(WatchKey watchKey, Path srcPath) throws Exception {
+			try {
+				if(watchKey!=null)
+					watchKey.cancel();
+//				// reset key and remove from set if directory no longer accessible
+//				boolean valid = key.reset();
+//				if (!valid) {
+//					if (trace) {
+//						System.out.println(key + " not valid");
+//						System.out.println("Will be removing " + keys.get(key) + " from key list");
+//					}
+//					keys.remove(key);
+	//
+//					// all directories are inaccessible
+//					if (keys.isEmpty()) {
+//						break;
+//					}
+//				}
+	//
+//				Collection<File> files = FileUtils.listFiles(artifactPath.toFile(), null, false);
+//				System.out.println(key + ":" + files.size());
+//				if(files.size() == 4) {
+//					System.out.println(artifactPath + " copy completed");
+//					
+//				}
+
+				CommandLineExecuterImpl clei = new CommandLineExecuterImpl();
+				clei.executeCommand("chmod -R 777 " + srcPath.toString(), false);
+				String artifactName = getArtifactName(srcPath);
+				final Path dest = Paths.get(opsDirLoc.toString(), artifactName, "mxf");
+				Files.createDirectories(dest);
+				Files.walkFileTree(srcPath, new SimpleFileVisitor<Path>() {
+					@Override
+					public FileVisitResult visitFile(Path file, BasicFileAttributes attrs) throws IOException {
+						Files.move(file, Paths.get(dest.toString(), file.getFileName().toString()), StandardCopyOption.ATOMIC_MOVE);
+						return FileVisitResult.CONTINUE;
+					}
+				});
+				Files.delete(srcPath);
+			} catch (IOException e) {
+				e.printStackTrace();
+				updateStatus(srcPath, Status.move_failed);
+			}
+		}
+	}
+
+    
+	
 	private void createCSVEntry(Path child) throws Exception {
 		String artifactName = getArtifactName(child);
 
@@ -377,47 +476,51 @@ public class DirectoryWatcherForCopiedFiles {
 			}
 		}
 	}
-	
-	private void updateStatus(Path child, Status status) throws Exception {
-		String artifactName = getArtifactName(child);
-		// Add only when artifact not there. For QC failure scenarios the file would come again...
-		for (String nthCSVEntry : csvEntries) {
-			if(nthCSVEntry.startsWith(artifactName + ", ")) {
-		        int index = csvEntries.indexOf(nthCSVEntry);
-
-				nthCSVEntry = nthCSVEntry.replaceAll(", [^,]*$", ", " + status.name());
-				System.out.println(artifactName + " status " + status.name());
-				csvEntries.set(index, nthCSVEntry);
-				// break; // Dont break so even if there are dupe entries they are updated
-			}
-		}
-		FileUtils.writeLines(csvFile, csvEntries);
-	}
     
-    
-	static void usage() {
-		System.err.println("usage: java DirectoryWatcher [waitTimeInSecs] <dirToBeWatched(\"/data/user/pgurumurthy/ingest\")> <opsDirLocation(\"/data/preprod/ops\")>");
-		System.exit(-1);
-	}
-
     private enum Status {
     	copying,
+    	copy_complete,
     	verifying,
-    	done,
+    	verified,
     	md5_mismatch,
     	moved,
     	move_failed,
     	failed;
     }
     
+	private void updateStatus(Path child, Status status) throws Exception {
+		logger.debug(String.format("%s %s", child, status.name()));
+		//updateLog(child + ":" + status.name());
+		
+//
+//		String artifactName = getArtifactName(child);
+//		// Add only when artifact not there. For QC failure scenarios the file would come again...
+//		for (String nthCSVEntry : csvEntries) {
+//			if(nthCSVEntry.startsWith(artifactName + ", ")) {
+//		        int index = csvEntries.indexOf(nthCSVEntry);
+//
+//				nthCSVEntry = nthCSVEntry.replaceAll(", [^,]*$", ", " + status.name());
+//				//System.out.println(artifactName + " status " + status.name());
+//				csvEntries.set(index, nthCSVEntry);
+//				// break; // Dont break so even if there are dupe entries they are updated
+//			}
+//		}
+//		FileUtils.writeLines(csvFile, csvEntries);
+	}
+
+	static void usage() {
+		System.err.println("usage: java DirectoryWatcher [waitTimeInSecs] <dirToBeWatched(\"/data/user/pgurumurthy/ingest\")> <opsDirLocation(\"/data/preprod/ops\")> noOfThreadsForChecksumVerification");
+		System.exit(-1);
+	}
+	
 	public static void main(String[] args) throws Exception {
 		// parse arguments
-		if (args.length == 0 || args.length > 3)
+		if (args.length == 0 || args.length > 4)
 			usage();
 
 		int dirArg = 0;
 		long waitTimes = 0L;
-		if (args.length == 3) {
+		if (args.length == 4) {
 			waitTimes = Long.parseLong(args[0]) * 1000;
 			dirArg++;
 		}
@@ -432,6 +535,9 @@ public class DirectoryWatcherForCopiedFiles {
 			FileUtils.write(csvFile, "artifact name, datetime copied, status", false);
 		
 		csvEntries = FileUtils.readLines(csvFile);
+		
+		int noOfThreads = Integer.parseInt(args[dirArg + 2]);
+		executor = new ThreadPoolExecutor(noOfThreads, noOfThreads, 0L, TimeUnit.MILLISECONDS, new LinkedBlockingQueue<Runnable>());
 
 		new DirectoryWatcherForCopiedFiles(dir, waitTimes).processEvents();
 	}