+++ ReleaseNotes.md
@@ -19,6 +19,7 @@
 > threadpoolexecutor.video-mam-update.corePoolSize=1
 > threadpoolexecutor.video-mam-update.maxPoolSize=1
 
+  * Deploy setpermissions script
 # Dwara App Version - 2.0.06
 # Dwara DB Version - 2.0.2
 ### New features